Extracted graph drawings.
